App/requirements.txt
PySide6>=6.0.0
pyinstaller>=6.0.0
pynput>=1.7.0
pantheon>=2.0.0
requests>=2.31.0
obs-websocket-py>=1.0
psutil>=5.9.0
